+++ core/src/box_constraints.hpp
@@ -5,22 +5,22 @@
 namespace aardvark {
 
 struct BoxConstraints {
-  int min_width;
-  int max_width;
-  int min_height;
-  int max_height;
+    float min_width;
+    float max_width;
+    float min_height;
+    float max_height;
 
-  // Makes constraints with same max size and zero min size
-  BoxConstraints make_loose();
+    // Makes constraints with same max size and zero min size
+    BoxConstraints make_loose();
 
-  // Checks whether min constraints is same as max constraints
-  bool is_tight();
+    // Checks whether min constraints is same as max constraints
+    bool is_tight();
 
-  // Returns max possible size
-  Size max_size();
+    // Returns max possible size
+    Size max_size();
 
-  // Makes tight or loose constraints from size
-  static BoxConstraints from_size(Size size, bool tight);
+    // Makes tight or loose constraints from size
+    static BoxConstraints from_size(Size size, bool tight);
 };
 
 } // namespace aardvark
